Finished initial culling
